Refactor useTabHandlers by domain
