+++ src/__tests__/renderer/hooks/tabs/internal/useFilePreviewTabHandlers.test.ts
@@ -1,0 +1,188 @@
+import { renderHook, act, cleanup } from '@testing-library/react';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { useFilePreviewTabHandlers } from '../../../../../renderer/hooks/tabs/internal/useFilePreviewTabHandlers';
+import { useModalStore } from '../../../../../renderer/stores/modalStore';
+import { useSettingsStore } from '../../../../../renderer/stores/settingsStore';
+import {
+	createMockAITab,
+	createMockFileTab,
+	getSession,
+	resetTabHandlerStores,
+	setupSession,
+} from './testUtils';
+
+describe('useFilePreviewTabHandlers', () => {
+	beforeEach(() => {
+		resetTabHandlerStores();
+	});
+
+	afterEach(() => {
+		cleanup();
+	});
+
+	it('opens a new file tab next to the active tab', () => {
+		setupSession({ aiTabs: [createMockAITab({ id: 'ai-1' })] });
+		const { result } = renderHook(() => useFilePreviewTabHandlers());
+
+		act(() => {
+			result.current.handleOpenFileTab({
+				path: '/repo/src/app.ts',
+				name: 'app.ts',
+				content: 'content',
+				lastModified: 55,
+			});
+		});
+
+		const session = getSession();
+		expect(session.filePreviewTabs[0]).toMatchObject({
+			path: '/repo/src/app.ts',
+			name: 'app',
+			extension: '.ts',
+			content: 'content',
+			lastModified: 55,
+		});
+		expect(session.activeFileTabId).toBe(session.filePreviewTabs[0].id);
+		expect(session.unifiedTabOrder.map((ref) => ref.type)).toEqual(['ai', 'file']);
+	});
+
+	it('updates and selects an existing file tab by path', () => {
+		const fileTab = createMockFileTab({
+			id: 'file-1',
+			path: '/repo/src/app.ts',
+			content: 'old',
+			isLoading: true,
+			loadRequestId: 'load-1',
+		});
+		setupSession({ filePreviewTabs: [fileTab] });
+		const { result } = renderHook(() => useFilePreviewTabHandlers());
+
+		act(() => {
+			result.current.handleOpenFileTab({
+				path: '/repo/src/app.ts',
+				name: 'app.ts',
+				content: 'new',
+				lastModified: 99,
+			});
+		});
+
+		expect(getSession().filePreviewTabs[0]).toMatchObject({
+			content: 'new',
+			lastModified: 99,
+			isLoading: false,
+			loadRequestId: undefined,
+		});
+		expect(getSession().activeFileTabId).toBe('file-1');
+	});
+
+	it('replaces the active file tab and truncates forward history', () => {
+		const fileTab = createMockFileTab({
+			id: 'file-1',
+			path: '/repo/b.ts',
+			name: 'b',
+			navigationHistory: [
+				{ path: '/repo/a.ts', name: 'a', scrollTop: 1 },
+				{ path: '/repo/b.ts', name: 'b', scrollTop: 2 },
+				{ path: '/repo/c.ts', name: 'c', scrollTop: 3 },
+			],
+			navigationIndex: 1,
+		});
+		setupSession({ filePreviewTabs: [fileTab], activeFileTabId: 'file-1' });
+		const { result } = renderHook(() => useFilePreviewTabHandlers());
+
+		act(() => {
+			result.current.handleOpenFileTab(
+				{ path: '/repo/d.ts', name: 'd.ts', content: 'd' },
+				{ openInNewTab: false }
+			);
+		});
+
+		expect(getSession().filePreviewTabs[0]).toMatchObject({
+			path: '/repo/d.ts',
+			name: 'd',
+			navigationIndex: 2,
+		});
+		expect(getSession().filePreviewTabs[0].navigationHistory).toEqual([
+			{ path: '/repo/a.ts', name: 'a', scrollTop: 1 },
+			{ path: '/repo/b.ts', name: 'b', scrollTop: 2 },
+			{ path: '/repo/d.ts', name: 'd', scrollTop: 0 },
+		]);
+	});
+
+	it('confirms before closing an edited file tab and cancels a loading read on confirm', () => {
+		const fileTab = createMockFileTab({
+			id: 'file-1',
+			name: 'app',
+			extension: '.ts',
+			editContent: 'dirty',
+			isLoading: true,
+			loadRequestId: 'load-1',
+		});
+		setupSession({ filePreviewTabs: [fileTab], activeFileTabId: 'file-1' });
+		const { result } = renderHook(() => useFilePreviewTabHandlers());
+
+		act(() => {
+			result.current.handleCloseFileTab('file-1');
+		});
+
+		const modal = useModalStore.getState().modals.get('confirm');
+		expect(modal?.data?.message).toContain('has unsaved changes');
+
+		act(() => {
+			modal?.data?.onConfirm();
+		});
+
+		expect(window.maestro.fs.cancelReadFile).toHaveBeenCalledWith('load-1');
+		expect(getSession().filePreviewTabs).toHaveLength(0);
+	});
+
+	it('auto-refreshes stale file content on selection when enabled', async () => {
+		const fileTab = createMockFileTab({
+			id: 'file-1',
+			path: '/repo/app.ts',
+			content: 'old',
+			lastModified: 1,
+		});
+		setupSession({ filePreviewTabs: [fileTab] });
+		useSettingsStore.setState({ fileTabAutoRefreshEnabled: true } as any);
+		vi.mocked(window.maestro.fs.stat).mockResolvedValue({
+			modifiedAt: new Date(5000).toISOString(),
+		} as any);
+		vi.mocked(window.maestro.fs.readFile).mockResolvedValue('fresh');
+		const { result } = renderHook(() => useFilePreviewTabHandlers());
+
+		await act(async () => {
+			await result.current.handleSelectFileTab('file-1');
+		});
+
+		expect(window.maestro.fs.readFile).toHaveBeenCalledWith('/repo/app.ts', undefined);
+		expect(getSession().activeFileTabId).toBe('file-1');
+		expect(getSession().filePreviewTabs[0].content).toBe('fresh');
+	});
+
+	it('navigates to an arbitrary file history index using the current SSH remote', async () => {
+		const fileTab = createMockFileTab({
+			id: 'file-1',
+			sshRemoteId: 'remote-1',
+			navigationHistory: [
+				{ path: '/repo/a.ts', name: 'a', scrollTop: 1 },
+				{ path: '/repo/b.ts', name: 'b', scrollTop: 2 },
+			],
+			navigationIndex: 0,
+		});
+		setupSession({ filePreviewTabs: [fileTab], activeFileTabId: 'file-1' });
+		vi.mocked(window.maestro.fs.readFile).mockResolvedValue('b-content');
+		const { result } = renderHook(() => useFilePreviewTabHandlers());
+
+		await act(async () => {
+			await result.current.handleFileTabNavigateToIndex(1);
+		});
+
+		expect(window.maestro.fs.readFile).toHaveBeenCalledWith('/repo/b.ts', 'remote-1');
+		expect(getSession().filePreviewTabs[0]).toMatchObject({
+			path: '/repo/b.ts',
+			content: 'b-content',
+			scrollTop: 2,
+			navigationIndex: 1,
+		});
+	});
+});
